server/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.3.1
gunicorn==22.0.0
python-dotenv==1.0.1
